Fix potential trailing -I flag
If Intl_INCLUDE_DIRS is empty, this becomes a trailing -I flag, which
eats up the subsequent arg (treating it as an include directory). This
can cause confusing compilation errors such as here:

https://hydra.nixos.org/build/151235004/nixlog/1

The error "you must specify either '-M' or '-MM'" happened because
the specified -MD flag followed a bare -I flag and was treated as
an include directory instead of a compiler option.