Drop ycollections classes
It was invented by me and shall be burried by me. It has been repurposed
for more than it was invented for BUT it's no longer used. And the
implementation is dangerous (for example, the insertion calls after
resizing all over the place should use placement new because the memory
is actually invalid).